+++ voltbro/math/math_ops.hpp
@@ -1,4 +1,5 @@
 #pragma once
+#include "voltbro/utils.hpp"
 #if defined(STM32G4) || defined(STM32_G)
 
 #include <cstdint>
@@ -19,33 +20,32 @@ extern "C" {
 #define arm_sin_f32 sinf
 #endif
 
-inline float fmaxf(float x, float y) {
+FORCE_INLINE float fmaxf(float x, float y) {
     /// Returns maximum of x, y ///
     return (((x) > (y)) ? (x) : (y));
 }
 
-inline float fminf(float x, float y) {
+FORCE_INLINE float fminf(float x, float y) {
     /// Returns minimum of x, y ///
     return (((x) < (y)) ? (x) : (y));
 }
 
-inline float fmaxf3(float x, float y, float z) {
+FORCE_INLINE float fmaxf3(float x, float y, float z) {
     /// Returns maximum of x, y, z ///
     return (x > y ? (x > z ? x : z) : (y > z ? y : z));
 }
 
-inline float fminf3(float x, float y, float z) {
+FORCE_INLINE float fminf3(float x, float y, float z) {
     /// Returns minimum of x, y, z ///
     return (x < y ? (x < z ? x : z) : (y < z ? y : z));
 }
 
-inline float roundf(float x) {
+FORCE_INLINE float roundf(float x) {
     /// Returns nearest integer ///
-
     return x < 0.0f ? ceilf(x - 0.5f) : floorf(x + 0.5f);
 }
 
-inline void limit_norm(float* x, float* y, float limit) {
+FORCE_INLINE void limit_norm(float* x, float* y, float limit) {
     /// Scales the lenght of vector (x, y) to be <= limit ///
     float norm = 0;
     arm_sqrt_f32(*x * *x + *y * *y, &norm);
@@ -56,22 +56,26 @@ inline void limit_norm(float* x, float* y, float limit) {
     }
 }
 
-inline int float_to_uint(float x, float x_min, float x_max, uint8_t bits) {
+FORCE_INLINE int float_to_uint(float x, float x_min, float x_max, uint8_t bits) {
     /// Converts a float to an unsigned int, given range and number of bits ///
     float span = x_max - x_min;
     float offset = x_min;
     return (int)((x - offset) * ((float)((1 << bits) - 1)) / span);
 }
 
-inline float uint_to_float(int x_int, float x_min, float x_max, uint8_t bits) {
+FORCE_INLINE float uint_to_float(int x_int, float x_min, float x_max, uint8_t bits) {
     /// converts unsigned int to float, given range and number of bits ///
     float span = x_max - x_min;
     float offset = x_min;
     return ((float)x_int) * span / ((float)((1 << bits) - 1)) + offset;
 }
 
-inline float mfmod(float x, float y) {
+FORCE_INLINE float mfmod(float x, float y) {
     return x - ((int)(x / y)) * y;
+}
+
+FORCE_INLINE void fast_abs(float* value) {
+    arm_abs_f32(value, value, 1);
 }
